refactor: extract analyze + build routes from run-now.ts (569 → 150 lines)
Move agent-analyzer and agent-builder routes into run-now-build.ts.
run-now.ts keeps only the POST /agents/:name/run execution route.
